+++ src/edit/InstanceForm.js
@@ -13,6 +13,7 @@ import {
   Row,
   Col,
   Button,
+  Icon,
   TextField,
   Select,
   Checkbox,
@@ -168,13 +169,6 @@ class InstanceForm extends React.Component {
 
     this.onToggleSection = this.onToggleSection.bind(this);
     this.cViewMetaData = this.props.stripes.connect(ViewMetaData);
-    this.paneHeaderDropdownItems = [
-      {
-        id: 'cancel-instance-edition',
-        label: <FormattedMessage id="ui-inventory.cancel" />,
-        onClick: props.onCancel,
-      }
-    ];
   }
 
   onToggleSection({ id }) {
@@ -196,6 +190,17 @@ class InstanceForm extends React.Component {
       <span data-test-header-title>
         <FormattedMessage id={titleTranslationKey} />
       </span>
+    );
+  }
+
+  getActionMenu = () => {
+    const { onCancel } = this.props;
+    return (
+      <Button buttonStyle="dropdownItem" id="cancel-instance-edition" onClick={onCancel}>
+        <Icon icon="hollowX">
+          <FormattedMessage id="ui-inventory.cancel" />
+        </Icon>
+      </Button>
     );
   }
 
@@ -277,7 +282,7 @@ class InstanceForm extends React.Component {
             onClose={onCancel}
             lastMenu={initialValues.id ? editInstanceLastMenu : addInstanceLastMenu}
             paneTitle={this.getPaneTitle()}
-            actionMenuItems={this.paneHeaderDropdownItems}
+            actionMenu={this.getActionMenu}
           >
             <div>
               <Headline size="large" tag="h3">{formatMsg({ id: 'ui-inventory.instanceRecord' })}</Headline>